Link to Godot v3.5 docs instead of the v3.4 ones
Before b0a7fa5 (Update to Godot v3.5.1, 2023-01-08), the Getting Started
guide told users to use Godot v3.4.1. Any links to the Godot manual
were, as a result, links to the manual for Godot’s 3.4 branch.

Now that we’re using Godot 3.5.x, it makes more sense to link to the
manual for Godot’s 3.5 branch.