Unset main scene
Scene 2 UWMF is not well suited for non-technical users. At the moment,
users have to fuss with Git just to do basic tasks like creating a level
or updating. It would be nice if users didn’t have to use Git if they
don’t want to.

This commit is a step in the direction. Eventually, I’m going to make it
so that this repo doesn’t contain a Godot project. Instead this repo
will just the contents of the project/wolf_editing_tools folder. Users
will create their own Godot project and either copy the
wolf_editing_tools folder into that project, or, if they want to,
include this repo in their Godot project by adding it as a submodule.